@@ -1,6 +1,6 @@
 {
   "name": "ozzy",
-  "version": "0.1.3",
+  "version": "0.2.0",
   "description": "A lightweight API interface",
   "main": "index.js",
   "repository": "https://github.com/duncangrubbs/ozzy.git",
@@ -9,17 +9,13 @@
   "scripts": {
     "build": "tsc -p .",
     "test": "echo \"ok\"",
-    "start": "node dist/example.js",
-    "coverage": "jest tests/*.test.js --env=jsdom --coverage && open ./coverage/lcov-report/index.html",
-    "prettier": "prettier --write --single-quote src/",
-    "prettier:check": "prettier --single-quote --check src/"
+    "example": "npm run build && node dist/example.js",
+    "prettier": "prettier --write --single-quote src/"
   },
   "devDependencies": {
-    "jest": "^26.4.2"
+    "prettier": "^3.2.5"
   },
   "dependencies": {
-    "node-fetch": "^2.6.6",
-    "prettier": "^2.4.1",
-    "typescript": "^4.4.4"
+    "typescript": "^5.4.5"
   }
 }
